fix clamp (#44)
This should fix the error reported in #43 

closes #43 
Bench: 2150799

+++ src/timemanager.cpp
@@ -11,7 +11,7 @@ Time optimumTime(int64_t avaiableTime, int inc, int ply, int mtg) {
     }
     time.optimum += inc / 2;
     if (time.optimum >= avaiableTime) {
-        time.optimum = std::clamp(time.optimum, 1LL, avaiableTime / 20);
+        time.optimum = std::clamp(time.optimum, (int64_t)1, avaiableTime / 20);
     }
 
     time.maximum = time.optimum;
